# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Poker API",
    "version": "0.1.0",
    "description": "Contract-first HTTP API for auth, economy, tournaments, lobby, profile, and admin."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Bearer token"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/economy/wallet": {
      "get": {
        "tags": [
          "economy"
        ],
        "operationId": "economyWallet",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Wallet balance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "bankrollCents": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "bankrollCents"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/economy/transactions": {
      "get": {
        "tags": [
          "economy"
        ],
        "operationId": "economyTransactions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  },
                  "required": [
                    "items"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/economy/buy-in": {
      "post": {
        "tags": [
          "economy"
        ],
        "operationId": "economyBuyIn",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tableId": {
                    "type": "string"
                  },
                  "amountCents": {
                    "type": "integer"
                  },
                  "externalRef": {
                    "type": "string"
                  }
                },
                "required": [
                  "tableId",
                  "amountCents"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Buy-in processed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "newTableBalance": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "success",
                    "newTableBalance"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/economy/cash-out": {
      "post": {
        "tags": [
          "economy"
        ],
        "operationId": "economyCashOut",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tableId": {
                    "type": "string"
                  },
                  "amountCents": {
                    "type": "integer"
                  },
                  "externalRef": {
                    "type": "string"
                  }
                },
                "required": [
                  "tableId",
                  "amountCents"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cash-out processed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments": {
      "get": {
        "tags": [
          "tournaments"
        ],
        "operationId": "tournamentsList",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tournament list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tournaments": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  },
                  "required": [
                    "tournaments"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tournaments"
        ],
        "operationId": "tournamentsCreate",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "entryFeeCents": {
                    "type": "integer"
                  },
                  "startTime": {
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "required": [
                  "name",
                  "entryFeeCents",
                  "startTime"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tournament created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "403": {
            "description": "Admin required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{id}": {
      "get": {
        "tags": [
          "tournaments"
        ],
        "operationId": "tournamentsGet",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tournament",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{id}/register": {
      "post": {
        "tags": [
          "tournaments"
        ],
        "operationId": "tournamentsRegister",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}